Object Uuid Provider: don't include outline as parent
Because there is never more than one outline shown at a time,
there is no need to include any outline in a uuidPath.
Also, appending the outline even for pages which have a classId would
break existing testing tools. It would be tolerable if it had a benefit,
but there is none.