ObjectUuidProvider: add support for dependent uuids
Some widgets like tabs or context-menu items are automatically created
and not specified by the developer. Hence, they don't have an
uuid or classId but still should be uniquely identifiable for testing
tools. They now use the uuid or classId of their main element with a
prefix. The prefix is necessary to avoid a conflict with the main
element (e.g. the uuid is not added twice to the DOM).