(!fixup) ObjectFactory: add ensureId decorator
Until now, every object created using scout.create() got a unique id
if no id was passed by the model, unless ensureUniqueId was set to
false when calling scout.create().

Normally, the id is either passed by the ui server to the model
adapters. Or set manually by the developer in the model files so a
widget can be found using Widget.widget('id'). If a developer did not
assign a specific id for a Scout JS widget, a unique id was assigned
automatically. This is sometimes useful for example when using the
function arrays.union().

But some objects just never need an id. Whether an id should be assigned
automatically can now be controlled with the decorator uniqueId.
Currently, the decorator is present on Widget, TableRow, TreeNode and
ModelAdapter.